backend/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-multipart>=0.0.9
httpx>=0.27.0
groq>=0.11.0
python-dotenv>=1.0.1
